// src/alice/types/yandex.ts
export enum YDevice {
	camera = "camera",
	cooking = "cooking",
	cooking_coffee_maker = "cooking.coffee_maker",
	cooking_kettle = "cooking.kettle",
	cooking_multicooker = "cooking.multicooker",
	dishwasher = "dishwasher",
	humidifier = "humidifier",
	iron = "iron",
	light = "light",
	light_ceiling = "light.ceiling",
	light_strip = "light.strip",
	media_device = "media_device",
	media_device_receiver = "media_device.receiver",
	media_device_tv = "media_device.tv",
	media_device_tv_box = "media_device.tv_box",
	openable = "openable",
	openable_curtain = "openable.curtain",
	openable_valve = "openable.valve",
	other = "other",
	pet_drinking_fountain = "pet_drinking_fountain",
	pet_feeder = "pet_feeder",
	purifier = "purifier",
	sensor = "sensor",
	sensor_button = "sensor.button",
	sensor_climate = "sensor.climate",
	sensor_gas = "sensor.gas",
	sensor_illumination = "sensor.illumination",
	sensor_motion = "sensor.motion",
	sensor_open = "sensor.open",
	sensor_smoke = "sensor.smoke",
	sensor_vibration = "sensor.vibration",
	sensor_water_leak = "sensor.water_leak",
	smart_meter = "smart_meter",
	smart_meter_cold_water = "smart_meter.cold_water",
	smart_meter_electricity = "smart_meter.electricity",
	smart_meter_gas = "smart_meter.gas",
	smart_meter_heat = "smart_meter.heat",
	smart_meter_hot_water = "smart_meter.hot_water",
	socket = "socket",
	switch = "switch",
	thermostat = "thermostat",
	thermostat_ac = "thermostat.ac",
	vacuum_cleaner = "vacuum_cleaner",
	ventilation = "ventilation",
	ventilation_fan = "ventilation.fan",
	washing_machine = "washing_machine",
}

export enum YCapability {
	on_off = "on_off",
	color_setting = "color_setting",
	video_stream = "video_stream",
	mode = "mode",
	range = "range",
	toggle = "toggle",
	float = "float",
	event = "event",
}

export enum YInstance {
	on = "on",
	hsv = "hsv",
	rgb = "rgb",
	temperature_k = "temperature_k",
	scene = "scene",
	get_stream = "get_stream",
	cleanup_mode = "cleanup_mode",
	coffee_mode = "coffee_mode",
	dishwashing = "dishwashing",
	fan_speed = "fan_speed",
	heat = "heat",
	input_source = "input_source",
	program = "program",
	swing = "swing",
	tea_mode = "tea_mode",
	thermostat = "thermostat",
	ventilation_mode = "ventilation_mode",
	work_speed = "work_speed",
	brightness = "brightness",
	channel = "channel",
	humidity = "humidity",
	open = "open",
	temperature = "temperature",
	volume = "volume",
	backlight = "backlight",
	controls_locked = "controls_locked",
	ionization = "ionization",
	keep_warm = "keep_warm",
	mute = "mute",
	oscillation = "oscillation",
	pause = "pause",
	amperage = "amperage",
	battery_level = "battery_level",
	co2_level = "co2_level",
	electricity_meter = "electricity_meter",
	food_level = "food_level",
	gas_meter = "gas_meter",
	heat_meter = "heat_meter",
	illumination = "illumination",
	meter = "meter",
	pm1_density = "pm1_density",
	pm2_5_density = "pm2.5_density",
	pm10_density = "pm10_density",
	power = "power",
	pressure = "pressure",
	tvoc = "tvoc",
	voltage = "voltage",
	water_level = "water_level",
	water_meter = "water_meter",
	vibration = "vibration",
	button = "button",
	motion = "motion",
	smoke = "smoke",
	gas = "gas",
	water_leak = "water_leak",
}

export enum YUnit {
	ampere = "ampere",
	cubic_meter = "cubic_meter",
	density_mcg_m3 = "density.mcg_m3",
	gigacalorie = "gigacalorie",
	illumination_lux = "illumination.lux",
	kilowatt_hour = "kilowatt_hour",
	percent = "percent",
	ppm = "ppm",
	pressure_atm = "pressure.atm",
	pressure_bar = "pressure.bar",
	pressure_mmhg = "pressure.mmhg",
	pressure_pascal = "pressure.pascal",
	temperature_celsius = "temperature.celsius",
	temperature_kelvin = "temperature.kelvin",
	volt = "volt",
	watt = "watt",
}

export enum YMode {
	wet_cleaning = "wet_cleaning",
	dry_cleaning = "dry_cleaning",
	mixed_cleaning = "mixed_cleaning",
	auto = "auto",
	eco = "eco",
	smart = "smart",
	turbo = "turbo",
	cool = "cool",
	dry = "dry",
	fan_only = "fan_only",
	heat = "heat",
	preheat = "preheat",
	high = "high",
	low = "low",
	medium = "medium",
	max = "max",
	min = "min",
	fast = "fast",
	slow = "slow",
	express = "express",
	normal = "normal",
	quiet = "quiet",
	horizontal = "horizontal",
	stationary = "stationary",
	vertical = "vertical",
	supply_air = "supply_air",
	extraction_air = "extraction_air",
	one = "one",
	two = "two",
	three = "three",
	four = "four",
	five = "five",
	six = "six",
	seven = "seven",
	eight = "eight",
	nine = "nine",
	ten = "ten",
	americano = "americano",
	cappuccino = "cappuccino",
	double_espresso = "double_espresso",
	espresso = "espresso",
	latte = "latte",
	black_tea = "black_tea",
	flower_tea = "flower_tea",
	green_tea = "green_tea",
	herbal_tea = "herbal_tea",
	oolong_tea = "oolong_tea",
	puerh_tea = "puerh_tea",
	red_tea = "red_tea",
	white_tea = "white_tea",
	glass = "glass",
	intensive = "intensive",
	pre_rinse = "pre_rinse",
	aspic = "aspic",
	baby_food = "baby_food",
	baking = "baking",
	bread = "bread",
	boiling = "boiling",
	cereals = "cereals",
	cheesecake = "cheesecake",
	deep_fryer = "deep_fryer",
	dessert = "dessert",
	fowl = "fowl",
	frying = "frying",
	macaroni = "macaroni",
	milk_porridge = "milk_porridge",
	multicooker = "multicooker",
	pasta = "pasta",
	pilaf = "pilaf",
	pizza = "pizza",
	sauce = "sauce",
	slow_cook = "slow_cook",
	soup = "soup",
	steam = "steam",
	stewing = "stewing",
	vacuum = "vacuum",
	yogurt = "yogurt",
}

export enum YEvent {
	tilt = "tilt",
	fall = "fall",
	vibration = "vibration",
	opened = "opened",
	closed = "closed",
	click = "click",
	double_click = "double_click",
	long_press = "long_press",
	detected = "detected",
	not_detected = "not_detected",
	empty = "empty",
	low = "low",
	normal = "normal",
	high = "high",
	dry = "dry",
	leak = "leak",
}